Make full Block class
I decided to use the knife class utilities instead of hump's, just because
it's a little simpler and I don't really need the flexibility of hump's.

This commit also includes a fix for some weird behavior you'd get when
moving the mouse outside of the screen while grabbing the block. Now, it
will remain correctly positioned in relation to the cursor.