Many changes, detailed below
Firstly, the addition of a Node class, from which Block now inherits.

Secondly, the addition of a *global* Game object, that contains most of
the interoperating parts, and handles dispatching events to Nodes (which
are registered with Game in the constructor).

Last and definitely least, I added more example blocks, so now there's one
of each color. After five seconds, the bottom row is deleted, as a display
of Node:delete. On a related note, there's now a little timer under the
FPS that lets you know how long the game's been running.