Change Node management system
Before this commit, every Node registered itself with the global Game
object. However, I don't think that was ideal. Therefore, now, every Node
is given a parent, with which it registers itself. The new <event>All()
methods will call the event methods on the Node it's called on, and all of
that Node's children.