Fix eventAll function
Before this, it did not append "All" to the name of the event when calling
it on children. This meant that it was working fine before, but when I
tried to add a "main" object (to hold both the grid and poly), it wasn't
calling the handlers for the children deeper than one level in.