Add semi-functioning transformation functions
Actually, they work pretty well. The issue is that certain rotations
result in Polies that appear especially offset, which just won't do. I'm
only committing this so that it'll be in the history, in case I ever come
back. In a moment, however, I will be implementing these functions using
matrix rotation instead, which should be considerably less finnicky to
work with.