Complete PolyDef rotation
This implementation converts the coordinates to a matrix, rotates the
matrix, and then converts it back to coordinates. It's probably not
efficient. But it doesn't matter, seeing as it will happen once, at the
beginning of the game, and there's just not that many definitions to
rotate in the first place.